requirements.txt
-i https://pypi.org/simple
aiohttp==3.6.2
alpaca-trade-api==0.48
alpha-vantage==2.2.0
async-timeout==3.0.1
asyncio-nats-client==0.10.0
attrs==19.3.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
deprecated==1.2.10
dnspython==1.16.0
idna==2.8
multidict==4.7.6
numpy==1.19.0
pandas==1.0.5
polygon-api-client==0.1.6
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.22.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.24.3
websocket-client==0.56.0
websockets==8.0.2
wrapt==1.12.1
yarl==1.4.2
